requirements.txt
absl-py==2.3.0
accelerate==1.7.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.12
aiosignal==1.3.2
alembic==1.16.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
audioread==3.0.1
backports-datetime-fromisoformat==2.0.3
bitsandbytes==0.45.3
braceexpand==0.1.7
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1
colorlog==6.9.0
contourpy==1.3.2
cuda-bindings==12.9.0
cuda-python==12.9.0
cycler==0.12.1
cytoolz==1.0.1
datasets==3.6.0
decorator==5.2.1
dill==0.3.8
Distance==0.1.3
dnspython==2.7.0
docopt==0.6.2
editdistance==0.8.1
einops==0.8.1
email_validator==2.2.0
exceptiongroup==1.3.0
executing==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
fiddle==0.3.0
filelock==3.13.1
fonttools==4.58.2
frozenlist==1.7.0
fsspec==2024.12.0
future==1.0.0
g2p-en==2.1.0
gitdb==4.0.12
GitPython==3.1.44
graphviz==0.20.3
greenlet==3.2.3
grpcio==1.73.0
h11==0.16.0
hf-xet==1.1.3
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.32.4
hydra-core==1.3.2
idna==3.10
inflect==7.5.0
intervaltree==3.1.0
ipython==8.37.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiwer==3.1.0
joblib==1.5.1
kaldi-python-io==1.2.2
kaldiio==2.18.1
kiwisolver==1.4.8
lazy_loader==0.4
Levenshtein==0.27.1
lhotse==1.30.3
libcst==1.8.0
librosa==0.11.0
lightning==2.4.0
lightning-utilities==0.14.3
lilcom==1.8.1
llvmlite==0.44.0
loguru==0.7.3
Mako==1.3.10
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==4.0.0
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapy==1.1.6
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.4.4
multiprocess==0.70.16
nemo-toolkit==2.3.1
networkx==3.3
nltk==3.9.1
num2words==0.5.14
numba==0.61.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-cufile-cu12==1.13.0.11
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.8.61
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
onnx==1.17.0
optuna==4.3.0
orjson==3.10.18
packaging==24.2
pandas==2.3.0
parso==0.8.4
peft==0.15.2
pexpect==4.9.0
pillow==11.2.1
plac==1.4.5
platformdirs==4.3.8
pooch==1.8.2
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==4.24.4
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyannote.core==5.0.0
pyannote.database==5.1.3
pyannote.metrics==3.2.1
pyarrow==20.0.0
pybind11==2.13.6
pycparser==2.22
pydantic==2.11.5
pydantic-extra-types==2.10.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.1
pyloudnorm==0.1.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytorch-lightning==2.5.1.post0
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.13.0
regex==2024.11.6
requests==2.32.4
resampy==0.4.3
rich==14.0.0
rich-toolkit==0.14.7
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
sacremoses==0.1.1
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.15.3
sentencepiece==0.2.0
sentry-sdk==2.29.1
setproctitle==1.3.6
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.13.1
sox==1.5.0
soxr==0.5.0.post1
SQLAlchemy==2.0.41
stack-data==0.6.3
starlette==0.46.2
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
termcolor==3.1.0
text-unidecode==1.3
texterrors==0.5.1
threadpoolctl==3.6.0
tokenizers==0.21.1
tomli==2.2.1
toolz==1.0.0
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchmetrics==1.7.2
torchvision==0.20.1+cu121
tqdm==4.67.1
traitlets==5.14.3
transformers==4.51.3
triton==3.1.0
typeguard==4.4.3
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
ujson==5.10.0
urllib3==2.4.0
uvicorn==0.34.3
uvloop==0.21.0
wandb==0.20.1
watchfiles==1.0.5
wcwidth==0.2.13
webdataset==0.2.111
websockets==15.0.1
Werkzeug==3.1.3
wget==3.2
wrapt==1.17.2
xxhash==3.5.0
yarl==1.20.1
